Change format of configuration file
Before this change we specified the mapping between certain characters
as dictionary. For example, one of the default mappings looked like
this: `o: ω◦ₒ`. We now use just use a list of characters to specify the
translation. The above mapping now looks like this: `oω◦ₒ`. The new
system is a little bit easier to use. It also shows that the first
character of the mapping is not in any way special.